requirements.txt
Flask==1.0.2
Flask-PyMongo==0.5.1
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
click==6.7
itsdangerous==0.24
pymongo==3.6.1
Flask-WTF==0.14.2
python-dotenv==0.8.2